# internal/api/testdata/openapi.golden.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TransitCiphertext": {
        "properties": {
          "ciphertext": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "ciphertext",
          "version"
        ],
        "type": "object"
      },
      "TransitDecryptRequest": {
        "properties": {
          "aad": {
            "format": "byte",
            "type": "string"
          },
          "ciphertext": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "ciphertext"
        ],
        "type": "object"
      },
      "TransitEncryptRequest": {
        "properties": {
          "aad": {
            "format": "byte",
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "plaintext": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "key",
          "plaintext"
        ],
        "type": "object"
      },
      "TransitHMAC": {
        "properties": {
          "hmac": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "hmac"
        ],
        "type": "object"
      },
      "TransitHMACRequest": {
        "properties": {
          "data": {
            "format": "byte",
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "data"
        ],
        "type": "object"
      },
      "TransitKey": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "name",
          "kind",
          "version"
        ],
        "type": "object"
      },
      "TransitKeyRequest": {
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "type": "object"
      },
      "TransitPlaintext": {
        "properties": {
          "plaintext": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "plaintext"
        ],
        "type": "object"
      },
      "TransitRewrapRequest": {
        "properties": {
          "aad": {
            "format": "byte",
            "type": "string"
          },
          "ciphertext": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "ciphertext"
        ],
        "type": "object"
      },
      "TransitRotateRequest": {
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "TransitSignRequest": {
        "properties": {
          "key": {
            "type": "string"
          },
          "message": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "key",
          "message"
        ],
        "type": "object"
      },
      "TransitSignature": {
        "properties": {
          "public_der": {
            "format": "byte",
            "type": "string"
          },
          "signature": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "signature",
          "public_der"
        ],
        "type": "object"
      },
      "TransitVerify": {
        "properties": {
          "valid": {
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "type": "object"
      },
      "TransitVerifyRequest": {
        "properties": {
          "message": {
            "format": "byte",
            "type": "string"
          },
          "public_der": {
            "format": "byte",
            "type": "string"
          },
          "signature": {
            "format": "byte",
            "type": "string"
          }
        },
        "required": [
          "message",
          "signature",
          "public_der"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "trstctl API token",
        "description": "Hashed API token resolved to a tenant-scoped principal with named trstctl permissions.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "description": "Verified browser session from OIDC, SAML, or LDAP; mutating requests also require the double-submit CSRF token.",
        "in": "cookie",
        "name": "trstctl_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Resource-oriented REST API for trstctl. Mutations require an Idempotency-Key; errors are RFC 7807 problem+json; lists use cursor pagination.",
    "title": "trstctl API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/transit/decrypt": {
      "post": {
        "operationId": "decryptTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitDecryptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitPlaintext"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Decrypt transit ciphertext",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/encrypt": {
      "post": {
        "operationId": "encryptTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitEncryptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitCiphertext"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Encrypt plaintext with a transit key",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/hmac": {
      "post": {
        "operationId": "hmacTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitHMACRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitHMAC"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Compute an HMAC with a transit key",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/keys": {
      "post": {
        "operationId": "createTransitKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Create a tenant-scoped transit key",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/keys/rotate": {
      "post": {
        "operationId": "rotateTransitKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitRotateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitKey"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Rotate a tenant-scoped transit key",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/rewrap": {
      "post": {
        "operationId": "rewrapTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitRewrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitCiphertext"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Re-encrypt transit ciphertext under the current key version",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/sign": {
      "post": {
        "operationId": "signTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitSignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitSignature"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Sign a message with a transit signing key",
        "x-trstctl-permission": "keys:write"
      }
    },
    "/api/v1/transit/verify": {
      "post": {
        "operationId": "verifyTransit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitVerify"
                }
              }
            },
            "description": "success"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "client error"
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "SessionCookie": []
          }
        ],
        "summary": "Verify a transit signature",
        "x-trstctl-permission": "keys:read"
      }
    },
...
  }
}